smartroom-api/api/requirements.txt
#API
asyncpg==0.25.0
uvicorn==0.17.4
fastapi==0.73.0
gunicorn==20.1.0
#for database 
databases==0.5.5
psycopg2==2.9.3
pydantic==1.9.0
SQLAlchemy==1.4.31
#for loading environment variables  
python-dotenv==0.19.2
paho-mqtt==1.6.1

